--- packages/graph-explorer/src/core/ConnectorProvider/types.ts
@@ -1,31 +1,0 @@
-import {
-  CountsByTypeRequest,
-  CountsByTypeResponse,
-  KeywordSearchRequest,
-  KeywordSearchResponse,
-  NeighborsCountRequest,
-  NeighborsCountResponse,
-  NeighborsRequest,
-  NeighborsResponse,
-  SchemaResponse,
-} from "../../connector/useGEFetchTypes";
-
-export type Explorer = {
-  fetchSchema: (options?: any) => Promise<SchemaResponse>;
-  fetchVertexCountsByType: (
-    req: CountsByTypeRequest,
-    options?: any
-  ) => Promise<CountsByTypeResponse>;
-  fetchNeighbors: (
-    req: NeighborsRequest,
-    options?: any
-  ) => Promise<NeighborsResponse>;
-  fetchNeighborsCount: (
-    req: NeighborsCountRequest,
-    options?: any
-  ) => Promise<NeighborsCountResponse>;
-  keywordSearch: (
-    req: KeywordSearchRequest,
-    options?: any
-  ) => Promise<KeywordSearchResponse>;
-};
